#411 separate report-logic and compare-logic
